Split comps code out & add proper attribution
Hif backend code for mapping comps to groups is
now split out into a separate file, to keep it
somewhat clean, especially as comps is currently
Fedora specific, while Hif and DNF are not intended to be so.

Additionally, the code for comps is properly attributed
to Helio and Kevin, since it is split out into separate files.